Fix warnings in shared connection code

+++ src/Connection.h
@@ -89,7 +89,7 @@ class Connection : public QObject {
 		void sendMessage(const QByteArray &qbaMsg);
 		void disconnectSocket(bool force=false);
 		void forceFlush();
-		int activityTime() const;
+		qint64 activityTime() const;
 		void resetActivityTime();
 
 		CryptState csCrypt;
